Refactoring

+++ src/elementoutput.cpp
@@ -135,7 +135,7 @@ void ElementOutput::updateValues()
 #ifdef USE_OPENMP
 #pragma omp parallel for
 #endif
-        for(int i = 0 ; i < m_geometries.size() ; i++)
+        for(int i = 0 ; i < (int)m_geometries.size() ; i++)
         {
           Element *element = m_component->modelInstance()->getElement(i);
           double value = element->atmosphericLWRadiation;
@@ -150,7 +150,7 @@ void ElementOutput::updateValues()
 #ifdef USE_OPENMP
 #pragma omp parallel for
 #endif
-        for(int i = 0 ; i < m_geometries.size() ; i++)
+        for(int i = 0 ; i < (int)m_geometries.size() ; i++)
         {
           Element *element = m_component->modelInstance()->getElement(i);
           double value = element->landCoverLWRadiation;
