+++ src/rhemodel.cpp
@@ -344,7 +344,7 @@ bool RHEModel::initializeElements(std::list<string> &errors)
 #ifdef USE_OPENMP
 #pragma omp parallel for
 #endif
-  for(int i = 0 ; i < m_elementJunctions.size()  ; i++)
+  for(int i = 0 ; i < (int)m_elementJunctions.size()  ; i++)
   {
     ElementJunction *elementJunction = m_elementJunctions[i];
     elementJunction->index = i;
@@ -370,7 +370,7 @@ bool RHEModel::initializeElements(std::list<string> &errors)
 #ifdef USE_OPENMP
 #pragma omp parallel for
 #endif
-  for(int i = 0; i < m_elements.size(); i++)
+  for(int i = 0; i < (int)m_elements.size(); i++)
   {
     Element *element = m_elements[i];
     element->index = i;
