+++ src/rhemodelio.cpp
@@ -411,7 +411,7 @@ bool RHEModel::initializeNetCDFOutputFile(list<string> &errors)
 #ifdef USE_OPENMP
 #pragma omp parallel for
 #endif
-    for (int i = 0; i < m_elements.size(); i++)
+    for (int i = 0; i < (int)m_elements.size(); i++)
     {
       Element *element = m_elements[i];
 
@@ -429,7 +429,7 @@ bool RHEModel::initializeNetCDFOutputFile(list<string> &errors)
     delete[] fromJunctions;
     delete[] toJunctions;
 
-    for (int i = 0; i < m_elements.size(); i++)
+    for (int i = 0; i < (int)m_elements.size(); i++)
     {
       delete[] elementIds[i];
     }
@@ -1798,7 +1798,7 @@ void RHEModel::writeNetCDFOutput()
 #ifdef USE_OPENMP
 #pragma omp parallel for
 #endif
-    for (int i = 0; i < m_elements.size(); i++)
+    for (int i = 0; i < (int)m_elements.size(); i++)
     {
       Element *element = m_elements[i];
       depth[i] = element->channelDepth;
